backend/requirements.txt
# Core
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0

# Database
pymongo==4.6.0
mongoengine==0.27.0

# HTTP Requests
requests==2.31.0

# Development
flake8==6.1.0
black==23.11.0
pytest==7.4.3

# Production
gunicorn==21.2.0

# Authentication (if needed)
PyJWT==2.8.0
Flask-JWT-Extended==4.5.3

# CORS (if your frontend is on a different domain)
Flask-Cors==4.0.0

# Validation
marshmallow==3.20.1